Added a new aoe type thick line
I added a new AOE tyle i call thick line. It's esentially a wall that is only spawns one controll token and is bound to the origin token which is the caster token. 
It's a messy addition that mostly just coppies the wall aoe sections. I was playing around so much with it so now i'm not even sure how much is diferent.
In the end i managed to get it to work.

I wanted a line that had more of a volume to it than the regular line provided. All i'll say is that i thank you for the comments spread throuought the code.
-Maks